refactor: Home과 MyPR 서버 컴포넌트 화
* 불필요한 props의 전달을 줄임
* useModalDispatch로 인한 클라이언트 컴포넌트화를 막음
