+++ src/app/(with-layout)/page.tsx
@@ -1,25 +1,11 @@
-import { RepositorySummaryType } from '@/__generated__/@types';
-import { apiApi } from '@/__generated__/Api/Api.api';
-import MyPR from '@/components/home/MyPR';
-import { sortRepositoriesByTracking } from '@/utils/sortRepositoriesByTracking';
+import { CloseAllModal } from '@/components/common/Modal/CloseAllModal';
+import { MyPR } from '@/components/home/MyPR';
 
-export default async function Home() {
-  let initialRepositoryList: RepositorySummaryType[];
-
-  try {
-    const { data: myRepositories } = await apiApi.getMyRepositories();
-    const repositoriesData = myRepositories.repositories ?? [];
-    const sortedRepositoriesByTracking = sortRepositoriesByTracking(repositoriesData);
-
-    const totalCount = repositoriesData.reduce((sum, r) => sum + (r.pullRequestCount || 0), 0);
-    initialRepositoryList = [{ id: 0, name: '전체', pullRequestCount: totalCount }, ...sortedRepositoriesByTracking];
-  } catch {
-    initialRepositoryList = [];
-  }
-
+export default function Home() {
   return (
     <div className={'mx-auto max-w-[1200px] px-[40px]'}>
-      <MyPR initialRepositoryList={initialRepositoryList} />
+      <MyPR />
+      <CloseAllModal />
     </div>
   );
 }
